+++ bfabric/docs/changelog.md
@@ -9,6 +9,10 @@ Minor breaking changes are still possible in `1.X.Y` but we try to announce them
 
 ## \[Unreleased\]
 
+### Added
+
+- `bfabric.experimental.webapp_integration_settings`
+
 ## \[1.16.1\] - 2025-12-15
 
 ### Fixed
@@ -126,6 +130,7 @@ Minor breaking changes are still possible in `1.X.Y` but we try to announce them
 - This is the last release supporting Python 3.9. From the next release onwards, Python 3.11 or higher will be required.
 - Flask is not a dependency anymore, as the dependency will be moved to bfabric-scripts.
 - Version upper bounds for dependencies have been defined.
+- More robust URL handling for token authentication.
 
 ### Fixed